+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+### Changed
+- Refactoring send flow
+
+### Fixed
+
 ## [3.7.8] - 2022-01-20
 
 ### Changed