+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Refactoring send flow
+- Jade: enabled testnet-liquid when in test mode
 
 ### Fixed