+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Select mnemonic length while restoring wallet
+- Access app settings screen from hardware wallet connecting screen
 
 ### Fixed